https://sourceforge.net/tracker/?func=detail&aid=2955366&group_id=243703&atid=1122994 Fix the optimize feature
git-svn-id: https://svn.code.sf.net/p/clapham/code/trunk@13 3b300aa7-9651-48eb-b63c-a8cf8e4f3f8d

+++ src/net/hydromatic/clapham/graph/Size.java
@@ -76,7 +76,7 @@ public class Size {
         this.width += width;
     }
 
-    public void incHeight(float height) {
+    public void incHeight(int height) {
         this.height += height;
     }
 }
